+++ src/main/java/com/study/service/ProfessorService.java
@@ -34,10 +34,13 @@ public class ProfessorService {
 		return professor;
 	}
 	
-	public void delete(final int id) {
-		repository.remove(id);
+	public void delete(final int id) throws Exception {
+		if(repository.containsKey(id)) {
+			repository.remove(id);
+		} else {
+			throw new Exception("Id não encontrado");
+		}
 	}
-	
 	
 }
